Add user's percentile to results
A new database table is created that tracks how many users got each number
of correct answers. I.e. how many got 0 correct, how many got 1
correct, ...

These values are updated when the quiz is finished and then echoed to the
user as a percentile that they are grouped with.